QuanLyHocLieuSo/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
keyboard==0.13.5
MouseInfo==0.1.3
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
PyAutoGUI==0.9.54
pycparser==2.23
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
wheel==0.45.1